refactor(web): refactor existing report components for improved data mart destination handling
