+++ apps/web/src/features/data-marts/reports/list/components/DestinationCard/index.ts
@@ -1,4 +1,4 @@
 export { DestinationCard } from './DestinationCard';
+export { AddReportButton } from './AddReportButton';
 export { ReportEditSheetRenderer } from './ReportEditSheetRenderer';
 export { ReportTableRenderer } from './ReportTableRenderer';
-export { AddReportButton } from './AddReportButton';
