--- apps/web/src/features/data-marts/reports/list/components/GoogleSheetsActionsCell/index.ts
@@ -1,1 +1,0 @@
-export { GoogleSheetsActionsCell } from './GoogleSheetsActionsCell.tsx';
