+++ apps/web/src/features/data-marts/reports/list/components/GoogleSheetsReportsTable/columns/columns.tsx
@@ -1,10 +1,9 @@
 import { type ColumnDef } from '@tanstack/react-table';
-import { GoogleSheetsActionsCell } from '../GoogleSheetsActionsCell';
-import { LookerStudioActionsCell } from '../LookerStudioActionsCell';
-import { StatusIcon } from '../StatusIcon';
-import { SortableHeader } from '../SortableHeader';
-import type { DataMartReport } from '../../../shared/model/types/data-mart-report.ts';
-import { ToggleColumnsHeader } from '../ToggleColumnsHeader';
+import { GoogleSheetsActionsCell } from '../GoogleSheetsActionsCell.tsx';
+import { StatusIcon } from '../../StatusIcon';
+import { SortableHeader } from '../SortableHeader.tsx';
+import type { DataMartReport } from '../../../../shared/model/types/data-mart-report';
+import { ToggleColumnsHeader } from '../ToggleColumnsHeader.tsx';
 import { ReportColumnKey } from './columnKeys.ts';
 import { ReportColumnLabels } from './columnLabels.ts';
 import RelativeTime from '@owox/ui/components/common/relative-time';
@@ -69,62 +68,6 @@ export const getGoogleSheetsColumns = ({
     header: ({ table }) => <ToggleColumnsHeader table={table} />,
     cell: ({ row }) => (
       <GoogleSheetsActionsCell
-        row={row}
-        onDeleteSuccess={onDeleteSuccess}
-        onEditReport={onEditReport}
-      />
-    ),
-    size: 80, // fixed width in pixels
-  },
-];
-
-export const getLookerStudioColumns = ({
-  onDeleteSuccess,
-  onEditReport,
-}: {
-  onDeleteSuccess?: () => void;
-  onEditReport?: (report: DataMartReport) => void;
-} = {}): (ColumnDef<DataMartReport> & {
-  meta?: { hidden?: boolean; title?: string };
-})[] => [
-  {
-    accessorKey: ReportColumnKey.TITLE,
-    header: ({ column }) => (
-      <SortableHeader column={column}>{ReportColumnLabels[ReportColumnKey.TITLE]}</SortableHeader>
-    ),
-    cell: ({ row }) => row.original.title,
-    enableColumnFilter: true,
-    size: 50, // responsive width in %
-  },
-  {
-    accessorKey: ReportColumnKey.LAST_RUN_DATE,
-    header: ({ column }) => <SortableHeader column={column}>Last Run Date</SortableHeader>,
-    cell: ({ row }) => {
-      const lastRunTimestamp = row.original.lastRunDate;
-      return (
-        <div className='text-muted-foreground text-sm'>
-          {lastRunTimestamp ? <RelativeTime date={new Date(lastRunTimestamp)} /> : 'Never run'}
-        </div>
-      );
-    },
-    size: 25, // responsive width in %
-  },
-  {
-    accessorKey: ReportColumnKey.LAST_RUN_STATUS,
-    header: ({ column }) => <SortableHeader column={column}>Last Run Status</SortableHeader>,
-    cell: ({ row }) =>
-      row.original.lastRunStatus ? (
-        <StatusIcon status={row.original.lastRunStatus} error={row.original.lastRunError} />
-      ) : (
-        <span className='text-muted-foreground text-sm'>&mdash;</span>
-      ),
-    size: 25, // responsive width in %
-  },
-  {
-    id: 'actions',
-    header: ({ table }) => <ToggleColumnsHeader table={table} />,
-    cell: ({ row }) => (
-      <LookerStudioActionsCell
         row={row}
         onDeleteSuccess={onDeleteSuccess}
         onEditReport={onEditReport}
