+++ apps/web/src/features/data-marts/reports/list/components/GoogleSheetsReportsTable/columns/index.ts
@@ -1,0 +1,2 @@
+export { getGoogleSheetsColumns, getAlignClass } from './columns';
+export type { Align } from './columns.tsx';
