+++ apps/web/src/features/data-marts/reports/list/components/GoogleSheetsReportsTable/index.ts
@@ -1,1 +1,5 @@
-export { GoogleSheetsReportsTable } from './GoogleSheetsReportsTable.tsx';
+export { GoogleSheetsReportsTable } from './GoogleSheetsReportsTable';
+export { getGoogleSheetsColumns, getAlignClass, type Align } from './columns';
+export { SortableHeader } from './SortableHeader';
+export { ToggleColumnsHeader } from './ToggleColumnsHeader';
+export { GoogleSheetsActionsCell } from './GoogleSheetsActionsCell';
