--- apps/web/src/features/data-marts/reports/list/components/LookerStudioActionsCell/index.ts
@@ -1,1 +1,0 @@
-export { LookerStudioActionsCell } from './LookerStudioActionsCell.tsx';
