--- apps/web/src/features/data-marts/reports/list/components/LookerStudioReportsTable/index.ts
@@ -1,1 +1,0 @@
-export { LookerStudioReportsTable } from './LookerStudioReportsTable.tsx';
