--- apps/web/src/features/data-marts/reports/list/components/SortableHeader/index.ts
@@ -1,1 +1,0 @@
-export { SortableHeader } from './SortableHeader.tsx';
