--- apps/web/src/features/data-marts/reports/list/components/TablePagination/index.ts
@@ -1,1 +1,0 @@
-export { TablePagination } from './TablePagination.tsx';
