--- apps/web/src/features/data-marts/reports/list/components/TableToolbar/index.ts
@@ -1,1 +1,0 @@
-export { TableToolbar } from './TableToolbar.tsx';
