--- apps/web/src/features/data-marts/reports/list/components/ToggleColumnsHeader/index.ts
@@ -1,1 +1,0 @@
-export { ToggleColumnsHeader } from './ToggleColumnsHeader.tsx';
