--- apps/web/src/features/data-marts/reports/list/components/columns/index.ts
@@ -1,2 +1,0 @@
-export { getGoogleSheetsColumns, getLookerStudioColumns, getAlignClass } from './columns.tsx';
-export type { Align } from './columns.tsx';
