+++ apps/web/src/features/data-marts/reports/list/components/index.ts
@@ -1,2 +1,5 @@
 export * from './GoogleSheetsReportsTable';
-export * from './LookerStudioReportsTable';
+export * from './LookerStudioReportCard';
+export * from './StatusIcon';
+export * from './EmptyDataMartDestinationsState';
+export * from './DestinationCard';
