+++ apps/web/src/features/data-marts/reports/list/model/hooks/index.ts
@@ -1,3 +1,4 @@
-export * from './useEditModal.ts';
-export * from './useTableFilter.ts';
-export * from './useColumnVisibility.ts';
+export * from './useEditModal';
+export * from './useColumnVisibility';
+export * from './useDestinationValidation';
+export * from './useReportModals';
