+++ apps/web/src/features/data-marts/reports/shared/components/index.ts
@@ -1,3 +1,1 @@
-export * from './DestinationCard';
 export * from './ReportHoverCard';
-export * from './EmptyDataMartDestinationsState';
