+++ apps/web/src/features/data-marts/reports/shared/model/hooks/index.ts
@@ -1,4 +1,2 @@
 export * from './useReport';
 export * from './useColumnVisibility';
-export * from './useReportModals';
-export * from './useDestinationValidation';
